cleaned the code and made only fgsm attack specific
